+++ src/sisl/physics/_matrix_dk.pyx
@@ -14,7 +14,8 @@ from ._matrix_phase import *
 from ._matrix_phase3 import *
 from ._phase import *
 
-__all__ = ["matrix_dk", "matrik_dk_nc", "matrik_dk_nc_diag", "matrik_dk_so"]
+__all__ = ["matrix_dk", "matrik_dk_nc", "matrik_dk_diag", "matrik_dk_so",
+    "matrix_dk_nambu"]
 
 
 def _phase_dk(gauge, M, sc, cnp.ndarray[floats_st] k, dtype):
@@ -36,6 +37,9 @@ def _phase_dk(gauge, M, sc, cnp.ndarray[floats_st] k, dtype):
         iRs = phase_rsc(sc, k, dtype).reshape(-1, 1)
         iRs = (1j * np.dot(sc.sc_off, sc.cell) * iRs).astype(dtype, copy=False)
         p_opt = 1
+
+
+    assert p_opt >= 0, "Not implemented"
 
     return p_opt, iRs
 
@@ -68,7 +72,8 @@ def matrix_dk_nc(gauge, M, sc, cnp.ndarray[floats_st] k, dtype, format):
     return d1.asformat(format), d2.asformat(format), d3.asformat(format)
 
 
-def matrix_dk_nc_diag(gauge, M, const ints_st idx, sc, cnp.ndarray[floats_st] k, dtype, format):
+def matrix_dk_diag(gauge, M, const ints_st idx, const ints_st per_row,
+                   sc, cnp.ndarray[floats_st] k, dtype, format):
     dtype = phase_dtype(k, M.dtype, dtype, True)
     p_opt, iRs = _phase_dk(gauge, M, sc, k, dtype)
 
@@ -80,14 +85,17 @@ def matrix_dk_nc_diag(gauge, M, const ints_st idx, sc, cnp.ndarray[floats_st] k,
     csr = M._csr
 
     if format in ("array", "matrix", "dense"):
-        x = _phase_array_nc_diag(csr.ptr, csr.ncol, csr.col, csr._D, idx, phx, p_opt)
-        y = _phase_array_nc_diag(csr.ptr, csr.ncol, csr.col, csr._D, idx, phy, p_opt)
-        z = _phase_array_nc_diag(csr.ptr, csr.ncol, csr.col, csr._D, idx, phz, p_opt)
+        x = _phase_array_diag(csr.ptr, csr.ncol, csr.col, csr._D, idx, phx, p_opt,
+        per_row)
+        y = _phase_array_diag(csr.ptr, csr.ncol, csr.col, csr._D, idx, phy, p_opt,
+        per_row)
+        z = _phase_array_diag(csr.ptr, csr.ncol, csr.col, csr._D, idx, phz, p_opt,
+        per_row)
 
     else:
-        x = _phase_csr_nc_diag(csr.ptr, csr.ncol, csr.col, csr._D, idx, phx, p_opt).asformat(format)
-        y = _phase_csr_nc_diag(csr.ptr, csr.ncol, csr.col, csr._D, idx, phy, p_opt).asformat(format)
-        z = _phase_csr_nc_diag(csr.ptr, csr.ncol, csr.col, csr._D, idx, phz, p_opt).asformat(format)
+        x = _phase_csr_diag(csr.ptr, csr.ncol, csr.col, csr._D, idx, phx, p_opt, per_row).asformat(format)
+        y = _phase_csr_diag(csr.ptr, csr.ncol, csr.col, csr._D, idx, phy, p_opt, per_row).asformat(format)
+        z = _phase_csr_diag(csr.ptr, csr.ncol, csr.col, csr._D, idx, phz, p_opt, per_row).asformat(format)
 
     return x, y, z
 
@@ -103,4 +111,18 @@ def matrix_dk_so(gauge, M, sc, cnp.ndarray[floats_st] k, dtype, format):
 
     # Default must be something else.
     d1, d2, d3 = _phase3_csr_so(csr.ptr, csr.ncol, csr.col, csr._D, iRs, p_opt)
+    return d1.asformat(format), d2.asformat(format), d3.asformat(format)
+
+
+def matrix_dk_nambu(gauge, M, sc, cnp.ndarray[floats_st] k, dtype, format):
+    dtype = phase_dtype(k, M.dtype, dtype, True)
+    p_opt, iRs = _phase_dk(gauge, M, sc, k, dtype)
+
+    csr = M._csr
+
+    if format in ("array", "matrix", "dense"):
+        return _phase3_array_nambu(csr.ptr, csr.ncol, csr.col, csr._D, iRs, p_opt)
+
+    # Default must be something else.
+    d1, d2, d3 = _phase3_csr_nambu(csr.ptr, csr.ncol, csr.col, csr._D, iRs, p_opt)
     return d1.asformat(format), d2.asformat(format), d3.asformat(format)