+++ src/sisl/physics/sparse.py
@@ -19,9 +19,21 @@ from sisl._internal import set_module
 from sisl.messages import warn
 from sisl.typing import AtomsIndex, GaugeType, KPoint
 
-from ._matrix_ddk import matrix_ddk, matrix_ddk_nc, matrix_ddk_nc_diag, matrix_ddk_so
-from ._matrix_dk import matrix_dk, matrix_dk_nc, matrix_dk_nc_diag, matrix_dk_so
-from ._matrix_k import matrix_k, matrix_k_nc, matrix_k_nc_diag, matrix_k_so
+from ._matrix_ddk import (
+    matrix_ddk,
+    matrix_ddk_diag,
+    matrix_ddk_nambu,
+    matrix_ddk_nc,
+    matrix_ddk_so,
+)
+from ._matrix_dk import (
+    matrix_dk,
+    matrix_dk_diag,
+    matrix_dk_nambu,
+    matrix_dk_nc,
+    matrix_dk_so,
+)
+from ._matrix_k import matrix_k, matrix_k_diag, matrix_k_nambu, matrix_k_nc, matrix_k_so
 from .spin import Spin
 
 __all__ = ["SparseOrbitalBZ", "SparseOrbitalBZSpin"]
@@ -675,8 +687,31 @@ class SparseOrbitalBZ(SparseOrbital):
            chosen gauge
         """
         k = _a.asarrayd(k).ravel()
-        return matrix_ddk_nc_diag(
-            gauge, self, self.S_idx, self.lattice, k, dtype, format
+        return matrix_ddk_diag(
+            gauge, self, self.S_idx, 2, self.lattice, k, dtype, format
+        )
+
+    def _ddSk_nambu(
+        self,
+        k: KPoint = (0, 0, 0),
+        dtype=None,
+        gauge: GaugeType = "cell",
+        format: str = "csr",
+    ):
+        r"""Overlap matrix in a `scipy.sparse.csr_matrix` at `k` for Nambu spin, differentiated with respect to `k`
+
+        Parameters
+        ----------
+        k : array_like, optional
+           k-point (default is Gamma point)
+        dtype : numpy.dtype, optional
+           default to `numpy.complex128`
+        gauge :
+           chosen gauge
+        """
+        k = _a.asarrayd(k).ravel()
+        return matrix_ddk_diag(
+            gauge, self, self.S_idx, 4, self.lattice, k, dtype, format
         )
 
     def eig(
@@ -835,6 +870,7 @@ class SparseOrbitalBZSpin(SparseOrbitalBZ):
                     2: Spin.POLARIZED,
                     4: Spin.NONCOLINEAR,
                     8: Spin.SPINORBIT,
+                    16: Spin.NAMBU,
                 }.get(dim)
         else:
             spin = kwargs.pop("spin")
@@ -910,6 +946,42 @@ class SparseOrbitalBZSpin(SparseOrbitalBZ):
             self.ddPk = self._ddPk_spin_orbit
             self.ddSk = self._ddSk_non_colinear
 
+        elif self.spin.is_nambu:
+            if self.dkind in ("f", "i"):
+                self.M11r = 0
+                self.M22r = 1
+                self.M12r = 2
+                self.M12i = 3
+                self.M11i = 4
+                self.M22i = 5
+                self.M21r = 6
+                self.M21i = 7
+                self.MSr = 8
+                self.MSi = 9
+                self.MT11r = 10
+                self.MT11i = 11
+                self.MT22r = 12
+                self.MT22i = 13
+                self.MT0r = 14
+                self.MT0i = 15
+            else:
+                self.M11 = 0
+                self.M22 = 1
+                self.M12 = 2
+                self.M21 = 3
+                self.MS = 4
+                self.MT11 = 5
+                self.MT22 = 6
+                self.MT0 = 7
+
+            # The overlap is the same as non-collinear
+            self.Pk = self._Pk_nambu
+            self.Sk = self._Sk_nambu
+            self.dPk = self._dPk_nambu
+            self.dSk = self._dSk_nambu
+            self.ddPk = self._ddPk_nambu
+            self.ddSk = self._ddSk_nambu
+
         if self.orthogonal:
             self.Sk = self._Sk_diagonal
 
@@ -975,7 +1047,102 @@ class SparseOrbitalBZSpin(SparseOrbitalBZ):
             dtype_cplx = dtype_real_to_complex(self.dtype)
 
             is_complex = self.dkind == "c"
-            if self.spin.is_spinorbit:
+            if self.spin.is_nambu:
+                if is_complex:
+                    nv = 8
+                    # Hermitian parameters
+                    # The input order is [uu, dd, ud, du]
+                    paramsH = [
+                        [
+                            # H^ee
+                            p[0].conjugate(),
+                            p[1].conjugate(),
+                            p[3].conjugate(),
+                            p[2].conjugate(),
+                            # delta, note the singlet
+                            -p[4].conjugate(),
+                            p[5].conjugate(),
+                            p[6].conjugate(),
+                            p[7].conjugate(),
+                            # because it is already off-diagonal
+                            *p[8:],
+                        ]
+                        for p in params
+                    ]
+                else:
+                    nv = 16
+                    # Hermitian parameters
+                    # The input order is [Ruu, Rdd, Rud, Iud, Iuu, Idd, Rdu, idu]
+                    #                    [ RS,  IS, RTu, ITu, RTd, ITd, RT0, IT0]
+                    # delta, note the singlet!
+                    paramsH = [
+                        [
+                            p[0],
+                            p[1],
+                            p[6],
+                            -p[7],
+                            -p[4],
+                            -p[5],
+                            p[2],
+                            -p[3],
+                            -p[8],
+                            p[9],
+                            p[10],
+                            -p[11],
+                            p[12],
+                            -p[13],
+                            p[14],
+                            -p[15],
+                            *p[16:],
+                        ]
+                        for p in params
+                    ]
+                if not self.orthogonal:
+                    nv += 1
+
+                # ensure we have correct number of values
+                assert all(len(p) == nv for p in params)
+
+                if R[0] <= 0.1001:  # no atom closer than 0.1001 Ang!
+                    # We check that the the parameters here is Hermitian
+                    p = params[0]
+                    if is_complex:
+                        Me = np.array([[p[0], p[2]], [p[3], p[1]]], dtype_cplx)
+                        # do Delta
+                        p = p[4:]
+                        Md = np.array(
+                            [[p[1], p[0] + p[3]], [-p[0] + p[3], p[2]]], dtype_cplx
+                        )
+                    else:
+                        Me = np.array(
+                            [
+                                [p[0] + 1j * p[4], p[2] + 1j * p[3]],
+                                [p[6] + 1j * p[7], p[1] + 1j * p[5]],
+                            ],
+                            dtype_cplx,
+                        )
+                        # do Delta
+                        p = p[8:]
+                        Md = np.array(
+                            [
+                                [p[2] + 1j * p[3], p[0] + p[6] + 1j * (p[1] + p[7])],
+                                [-p[0] + p[6] + 1j * (-p[1] + p[7]), p[4] + 1j * p[5]],
+                            ],
+                            dtype_cplx,
+                        )
+                    if not np.allclose(Me, Me.T.conjugate()):
+                        warn(
+                            f"{self.__class__.__name__}.create_construct is NOT "
+                            "Hermitian for M^e on-site terms. This is your responsibility! "
+                            "The code will continue silently, be AWARE!"
+                        )
+                    if not np.allclose(Md, Md.T.conjugate()):
+                        warn(
+                            f"{self.__class__.__name__}.create_construct is NOT "
+                            "Hermitian for Delta on-site terms. This is your responsibility! "
+                            "The code will continue silently, be AWARE!"
+                        )
+            elif self.spin.is_spinorbit:
                 if is_complex:
                     nv = 4
                     # Hermitian parameters
@@ -1092,7 +1259,8 @@ class SparseOrbitalBZSpin(SparseOrbitalBZ):
             Spin.POLARIZED: "polarized",
             Spin.NONCOLINEAR: "noncolinear",
             Spin.SPINORBIT: "spinorbit",
-        }.get(self.spin._kind, f"unkown({self.spin._kind})")
+            Spin.NAMBU: "nambu",
+        }.get(self.spin.kind, f"unkown({self.spin.kind})")
         return f"<{self.__module__}.{self.__class__.__name__} na={g.na}, no={g.no}, nsc={g.nsc}, dim={self.dim}, nnz={self.nnz}, spin={spin}>"
 
     def _Pk_unpolarized(
@@ -1180,6 +1348,27 @@ class SparseOrbitalBZSpin(SparseOrbitalBZ):
         k = _a.asarrayd(k).ravel()
         return matrix_k_so(gauge, self, self.lattice, k, dtype, format)
 
+    def _Pk_nambu(
+        self,
+        k: KPoint = (0, 0, 0),
+        dtype=None,
+        gauge: GaugeType = "cell",
+        format: str = "csr",
+    ):
+        r"""Sparse matrix (`scipy.sparse.csr_matrix`) at `k` for a Nambu system
+
+        Parameters
+        ----------
+        k : array_like, optional
+           k-point (default is Gamma point)
+        dtype : numpy.dtype, optional
+           default to `numpy.complex128`
+        gauge :
+           chosen gauge
+        """
+        k = _a.asarrayd(k).ravel()
+        return matrix_k_nambu(gauge, self, self.lattice, k, dtype, format)
+
     def _dPk_unpolarized(
         self,
         k: KPoint = (0, 0, 0),
@@ -1251,7 +1440,7 @@ class SparseOrbitalBZSpin(SparseOrbitalBZ):
         gauge: GaugeType = "cell",
         format: str = "csr",
     ):
-        r"""Tuple of sparse matrix (`scipy.sparse.csr_matrix`) at `k` for a non-collinear system, differentiated with respect to `k`
+        r"""Tuple of sparse matrix (`scipy.sparse.csr_matrix`) at `k` for a spin-orbit system, differentiated with respect to `k`
 
         Parameters
         ----------
@@ -1264,6 +1453,27 @@ class SparseOrbitalBZSpin(SparseOrbitalBZ):
         """
         k = _a.asarrayd(k).ravel()
         return matrix_dk_so(gauge, self, self.lattice, k, dtype, format)
+
+    def _dPk_nambu(
+        self,
+        k: KPoint = (0, 0, 0),
+        dtype=None,
+        gauge: GaugeType = "cell",
+        format: str = "csr",
+    ):
+        r"""Tuple of sparse matrix (`scipy.sparse.csr_matrix`) at `k` for a Nambu spin system, differentiated with respect to `k`
+
+        Parameters
+        ----------
+        k : array_like, optional
+           k-point (default is Gamma point)
+        dtype : numpy.dtype, optional
+           default to `numpy.complex128`
+        gauge :
+           chosen gauge
+        """
+        k = _a.asarrayd(k).ravel()
+        return matrix_dk_nambu(gauge, self, self.lattice, k, dtype, format)
 
     def _ddPk_non_colinear(
         self,
@@ -1293,7 +1503,7 @@ class SparseOrbitalBZSpin(SparseOrbitalBZ):
         gauge: GaugeType = "cell",
         format: str = "csr",
     ):
-        r"""Tuple of sparse matrix (`scipy.sparse.csr_matrix`) at `k` for a non-collinear system, differentiated with respect to `k`
+        r"""Tuple of sparse matrix (`scipy.sparse.csr_matrix`) at `k` for a spin-orbit system, differentiated with respect to `k`
 
         Parameters
         ----------
@@ -1306,6 +1516,27 @@ class SparseOrbitalBZSpin(SparseOrbitalBZ):
         """
         k = _a.asarrayd(k).ravel()
         return matrix_ddk_so(gauge, self, self.lattice, k, dtype, format)
+
+    def _ddPk_nambu(
+        self,
+        k: KPoint = (0, 0, 0),
+        dtype=None,
+        gauge: GaugeType = "cell",
+        format: str = "csr",
+    ):
+        r"""Tuple of sparse matrix (`scipy.sparse.csr_matrix`) at `k` for a Nambu system, differentiated with respect to `k`
+
+        Parameters
+        ----------
+        k : array_like, optional
+           k-point (default is Gamma point)
+        dtype : numpy.dtype, optional
+           default to `numpy.complex128`
+        gauge :
+           chosen gauge
+        """
+        k = _a.asarrayd(k).ravel()
+        return matrix_ddk_nambu(gauge, self, self.lattice, k, dtype, format)
 
     def _Sk(
         self,
@@ -1346,7 +1577,28 @@ class SparseOrbitalBZSpin(SparseOrbitalBZ):
            chosen gauge
         """
         k = _a.asarrayd(k).ravel()
-        return matrix_k_nc_diag(gauge, self, self.S_idx, self.lattice, k, dtype, format)
+        return matrix_k_diag(gauge, self, self.S_idx, 2, self.lattice, k, dtype, format)
+
+    def _Sk_nambu(
+        self,
+        k: KPoint = (0, 0, 0),
+        dtype=None,
+        gauge: GaugeType = "cell",
+        format: str = "csr",
+    ):
+        r"""Overlap matrix (`scipy.sparse.csr_matrix`) at `k` for a Nambu system
+
+        Parameters
+        ----------
+        k : array_like, optional
+           k-point (default is Gamma point)
+        dtype : numpy.dtype, optional
+           default to `numpy.complex128`
+        gauge :
+           chosen gauge
+        """
+        k = _a.asarrayd(k).ravel()
+        return matrix_k_diag(gauge, self, self.S_idx, 4, self.lattice, k, dtype, format)
 
     def _dSk_non_colinear(
         self,
@@ -1367,8 +1619,31 @@ class SparseOrbitalBZSpin(SparseOrbitalBZ):
            chosen gauge
         """
         k = _a.asarrayd(k).ravel()
-        return matrix_dk_nc_diag(
-            gauge, self, self.S_idx, self.lattice, k, dtype, format
+        return matrix_dk_diag(
+            gauge, self, self.S_idx, 2, self.lattice, k, dtype, format
+        )
+
+    def _dSk_nambu(
+        self,
+        k: KPoint = (0, 0, 0),
+        dtype=None,
+        gauge: GaugeType = "cell",
+        format: str = "csr",
+    ):
+        r"""Overlap matrix (`scipy.sparse.csr_matrix`) at `k` for a Nambu system
+
+        Parameters
+        ----------
+        k : array_like, optional
+           k-point (default is Gamma point)
+        dtype : numpy.dtype, optional
+           default to `numpy.complex128`
+        gauge :
+           chosen gauge
+        """
+        k = _a.asarrayd(k).ravel()
+        return matrix_dk_diag(
+            gauge, self, self.S_idx, 4, self.lattice, k, dtype, format
         )
 
     def eig(
@@ -1508,7 +1783,40 @@ class SparseOrbitalBZSpin(SparseOrbitalBZ):
         sp = self.spin
         D = new._csr._D
 
-        if sp.is_spinorbit:
+        if sp.is_nambu:
+            if hermitian and spin:
+                # conjugate the imaginary value and transpose spin-box
+                if self.dkind in ("f", "i"):
+                    # imaginary components (including transposing)
+                    #    12,11,22,21
+                    D[:, [3, 4, 5, 7]] = -D[:, [7, 4, 5, 3]]
+                    # R12 <-> R21
+                    D[:, [2, 6]] = D[:, [6, 2]]
+                    # real S, otherwise imaginary components of Delta
+                    D[:, [8, 11, 13, 15]] = -D[:, [8, 11, 13, 15]]
+                else:
+                    D[:, [0, 1, 2, 3]] = np.conj(D[:, [0, 1, 3, 2]])
+                    # delta values
+                    D[:, 4:8] = np.conj(D[:, 4:8])
+                    D[:, 4] = -D[:, 4]
+            elif hermitian:
+                # conjugate the imaginary value
+                if self.dkind in ("f", "i"):
+                    # imaginary components
+                    #    12,11,22,21
+                    D[:, [3, 4, 5, 7, 9, 11, 13, 15]] *= -1.0
+                else:
+                    D[:, :] = np.conj(D[:, :])
+            elif spin:
+                # transpose spin-box, 12 <-> 21
+                if self.dkind in ("f", "i"):
+                    D[:, [2, 3, 6, 7]] = D[:, [6, 7, 2, 3]]
+                    D[:, [8, 9]] = -D[:, [8, 9]]
+                else:
+                    D[:, [2, 3]] = D[:, [3, 2]]
+                    D[:, 4] = -D[:, 4]
+
+        elif sp.is_spinorbit:
             if hermitian and spin:
                 # conjugate the imaginary value and transpose spin-box
                 if self.dkind in ("f", "i"):
@@ -1569,7 +1877,10 @@ class SparseOrbitalBZSpin(SparseOrbitalBZ):
         D = new._csr._D
 
         # Apply Pauli-Y on the left and right of each spin-box
-        if sp.is_spinorbit:
+        if sp.is_nambu:
+            raise NotImplementedError
+
+        elif sp.is_spinorbit:
             if self.dkind in ("f", "i"):
                 # [R11, R22, R12, I12, I11, I22, R21, I21]
                 # [R11, R22] = [R22, R11]
@@ -1579,13 +1890,19 @@ class SparseOrbitalBZSpin(SparseOrbitalBZ):
                 # [R12, R21] = -[R21, R12] (Y @ Y)
                 D[:, [4, 5, 2, 6]] = -D[:, [5, 4, 6, 2]]
             else:
-                raise NotImplementedError
+                # [R11, R22, R12, I12, I11, I22, R21, I21]
+                # [11, 22] = [22, 11]^*
+                D[:, [0, 1]] = np.conj(D[:, [1, 0]])
+                # [12, 21] = -[21, 12]^* (Y @ Y)
+                D[:, [2, 3]] = -np.conj(D[:, [3, 2]])
+
         elif sp.is_noncolinear:
             if self.dkind in ("f", "i"):
                 # [R11, R22, R12, I12]
                 D[:, 2] = -D[:, 2]
             else:
-                raise NotImplementedError
+                # [R11, R22, 12]
+                D[:, 2] = -np.conj(D[:, 2])
 
         return new